Allow run mixed streams from CLI
