+++ application/holder/src/service/cli/index.ts
@@ -1,9 +1,6 @@
 export { Action as OpenFile } from './open';
 export { Action as Search } from './search';
-export { Action as Stdout } from './stdout';
-export { Action as Udp } from './udp';
-export { Action as Tcp } from './tcp';
-export { Action as Serial } from './serial';
+export { Action as Stream } from './stream';
 export { Action as Parser } from './parser';
 export { Type } from './action';
 export { CLIAction } from './action';
